Use smoother alpha and reduce laser fire rate
NOTE: `alpha` seems to depend on start `dt` accumulated.
This varies based on each game instance's startup conditions.
e.g., sometimes alpha could be .6111 or .8234 and fluctuate from there,
but does not deviate a lot from the starting alpha value.

We need to look into how this all works and make it consistent, no
matter the device running the game.